+++ android/src/main/java/asia/ivity/mediainfo/VideoUtils.java
@@ -5,7 +5,6 @@ import android.media.MediaFormat;
 import android.media.MediaMetadataRetriever;
 import android.os.Build;
 import android.util.Log;
-
 import java.io.File;
 
 class VideoUtils {
@@ -24,13 +23,18 @@ class VideoUtils {
       int height =
           Integer.parseInt(
               retriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_VIDEO_HEIGHT));
-      int rotation = Integer.valueOf(retriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_VIDEO_ROTATION));
 
-      // Switch the width/height if video was taken in portrait mode
-      if (rotation == 90 || rotation == 270) {
-        int temp = width;
-        width = height;
-        height = temp;
+      if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.JELLY_BEAN_MR1) {
+        int rotation = Integer
+            .valueOf(retriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_VIDEO_ROTATION));
+
+        // Switch the width/height if video was taken in portrait mode
+        if (rotation == 90 || rotation == 270) {
+          int temp = width;
+          //noinspection SuspiciousNameCombination
+          width = height;
+          height = temp;
+        }
       }
 
       float frameRate;
@@ -58,7 +62,6 @@ class VideoUtils {
           retriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_MIMETYPE);
 
       return new VideoDetail(width, height, frameRate, durationMs, numTracks, mimeType);
-
     } catch (Throwable e) {
       Log.e(TAG, file.getAbsolutePath(), e);
       return null;