+++ ios/Classes/IVMediaInfoPlugin.h
@@ -1,0 +1,4 @@
+#import <Flutter/Flutter.h>
+
+@interface IVMediaInfoPlugin : NSObject<FlutterPlugin>
+@end