+++ ios/Classes/MediaInfoPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface MediaInfoPlugin : NSObject<FlutterPlugin>
-@end