+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## next
 
+* iOS: Remove unused logging
 * iOS: Migrate iOS project
+* Adds basic method plugin tests on Dart side
 
 ## 0.5.2