+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## next
 
+* Android: Video rotation can not be detected on Android < 17.
 * Support the v2 Android embedder.
 * Bump a few example dependency versions
 * iOS: Remove unused logging